+++ src/transform/function/transform_handler_fn.py
@@ -1,5 +1,5 @@
 from .utils.get_latest_files import get_latest_files
-from .utils.list_s3 import list_s3
+from .utils.list_s3 import ingested_data_retrival
 from .utils.s3_client import s3_client
 import logging
 import os
@@ -9,7 +9,7 @@ def transform_handler(event, context):
     """
     Is invoked when triggered by an event
 
-    Invokes s3_client, files_dict and list_s3.
+    Invokes s3_client, files_dict and ingested_data_retrival.
 
         Parameters:
             event (event object): json object - converted to empty dict
@@ -30,7 +30,7 @@ def transform_handler(event, context):
     
     files_dict = get_latest_files(s3_client, bucket_name, logger)
     
-    list_s3(s3_client, files_dict, logger, bucket_name)
+    ingested_data_retrival(s3_client, files_dict, logger, bucket_name)
     
     return {"statusCode": 200, "body": "Data transformation complete"}
